se agrego la funcion que develve todos las notificaciones generales al repositorio e interfaz de Notifications y se agrego la funcion que envia el correo a todos los usuarios de las notificaciones generales

+++ Mhotivo.Data/Entities/NotificationType.cs
@@ -8,7 +8,7 @@ namespace Mhotivo.Data.Entities
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public long NotificationTypeId { get; set; }
+        public int NotificationTypeId { get; set; }
 
         public string TypeDescription { get; set; }
